+++ src/controlers/facture.js
@@ -83,10 +83,10 @@ class FactureController {
     }
 
     async delete(req, res) {
-        // const user = req.user;
-        // if(user.type > 2 || user.user_id != req.params.id) {
-        //     return res.status(403).json({ error: true, message: "Forbiden" });
-        // }
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
 
         prisma.facture.delete({
             where: { id: req.params.id }
@@ -96,10 +96,10 @@ class FactureController {
     }
 
     update(req, res) {
-        // const user = req.user;
-        // if(user.type > 2 || user.user_id != req.params.id) {
-        //     return res.status(403).json({ error: true, message: "Forbiden" });
-        // }
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
 
         const id = req.params.id;
         const params = req.body;
@@ -117,6 +117,11 @@ class FactureController {
         const facture = res.params.id;
         const params = req.body;
 
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
+
         if(params?.id !== undefined) delete params.id;
         Object.defineProperty(params, 'factureId', { value: facture, enumerable: true, writable: true });
 
@@ -127,6 +132,11 @@ class FactureController {
 
     deleteLigne(req, res) {
         const id = res.params.ligneId;
+
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
         
         prisma.ligneFacture.delete({
             where: { id }
